requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.2
python-dotenv==1.0.1
sqlparse==0.5.1
asgiref==3.8.1
psycopg2-binary==2.9.9
tzdata==2024.1
django-filter==24.3
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0